Move configuration routes over
