+++ modules/ept-configurations/index.js
@@ -1,6 +1,6 @@
 var path = require('path');
 var db = require(path.normalize(__dirname + '/db'));
-// var routes = require(path.normalize(__dirname + '/routes'));
+var routes = require(path.normalize(__dirname + '/routes'));
 // var authorization = require(path.normalize(__dirname + '/authorization'));
 // var permissions = require(path.normalize(__dirname + '/permissions'));
 
@@ -8,6 +8,6 @@ module.exports =  {
   name: 'configurations',
   // authorization: authorization,
   // permissions: permissions,
-  // routes: routes,
+  routes: routes,
   db: db
 };
