+++ modules/ept-configurations/routes/index.js
@@ -1,0 +1,6 @@
+var path = require('path');
+
+module.exports = [
+  require(path.normalize(__dirname + '/get')),
+  require(path.normalize(__dirname + '/update'))
+];
